Move `FeedUIIntegrationTests` test helper methods to it's own files
