+++ EssentialFeed/EssentialFeediOSTests/Feed UI/Helpers/FeedUIIntegrationTests+LoaderSpy.swift
@@ -1,0 +1,69 @@
+//
+//  FeedUIIntegrationTests+LoaderSpy.swift
+//  EssentialFeediOSTests
+//
+//  Created by Hitesh on 09/02/22.
+//
+
+import XCTest
+import EssentialFeed
+import EssentialFeediOS
+
+extension FeedUIIntegrationTests {
+    
+    class LoaderSpy: FeedLoader, FeedImageDataLoader {
+        
+        // MARK: - FeedLoader
+        
+        private var feedRequests : [(FeedLoader.Result) -> Void] = []
+        
+        var loadFeedCallCount: Int {
+            return feedRequests.count
+        }
+                
+        func load(completion: @escaping (FeedLoader.Result) -> Void) {
+            self.feedRequests.append(completion)
+        }
+        
+        func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
+            self.feedRequests[index](.success(feed))
+        }
+        
+        func completeFeedLoadingWithError(at index: Int = 0) {
+            let error = NSError(domain: "an error", code: 0)
+            self.feedRequests[index](.failure(error))
+        }
+        
+        // MARK: - FeedImageDataLoader
+        
+        private struct TaskSpy: FeedImageDataLoaderTask {
+            let cancelCallback: () -> Void
+            func cancel() {
+                cancelCallback()
+            }
+        }
+        
+        private (set) var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)] ()
+        
+        var loadedImageURLs: [URL] {
+            return imageRequests.map { $0.url }
+        }
+        
+        private (set) var cancelledImageURLs = [URL]()
+        
+        func loadImageData(from url: URL, completion:@escaping(FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+            imageRequests.append((url, completion))
+            return TaskSpy { [weak self] in self?.cancelledImageURLs.append(url) }
+        }
+        
+        func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
+            imageRequests[index].completion(.success(imageData))
+        }
+        
+        func completeImageLoadingWithError(at index: Int = 0) {
+            let error = NSError(domain: "an error", code: 0)
+            imageRequests[index].completion(.failure(error))
+        }
+    }
+    
+}
