+++ EssentialFeed/EssentialFeediOSTests/Feed UI/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,18 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Hitesh on 09/02/22.
+//
+
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        self.allTargets.forEach { target in
+            self.actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
+                (target as NSObject).perform(Selector($0))
+            }
+        }
+    }
+}
